backend/requirements.txt
Flask>=0.12.2
Flask-API>=1.0
Werkzeug>=0.14.1
gunicorn>=19.7.1
requests>=2.18.4
Flask-Testing>=0.7.1
mock>=2.0.0
requests-mock>=1.4.0
raven[flask]>=6.5.0
haversine>=0.4.5
jsonpickle>=0.9.6
behave>=1.2.5
pyhamcrest>=1.9.0
WebTest